IRanges 2.41.3
New implementations of the viewMins(), viewMaxs(), viewSums(), viewMeans(),
viewWhichMins(), viewWhichMaxs(), viewRangeMins(), and viewRangeMaxs()
methods for RleViews objects.

This is a complete rewrite (including of the C code behind these methods)
that addresses many long-standing (and unnoticed-for-years) bugs in the
previous implementation. This new implementation can also be significantly
faster than the previous one e.g. 5x faster or more for RleViews objects
with millions of views.