IRanges 2.43.7: Define Views() methods for integer and numeric vectors
These methods are already defined in the XVector package which is listed
in Suggests. The new methods in IRanges delegate to the methods in XVector.
This way people can do something like Views(101:110, start=5:1, end=5:9)
right after loading IRanges, and without having to explicitely load XVector,
and it'll just work.

This partly addresses issue #63.